src/requirements.txt
Authlib==1.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.4
cryptography==37.0.4
dnspython==2.2.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.0
Flask-PyMongo==2.3.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.4.0
pymongo==4.1.1
python-dateutil==1.5
python-dotenv==0.20.0
requests==2.28.1
six==1.16.0
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.0
